lead_management/requirements.txt
django
djangorestframework
psycopg2-binary
python-dotenv

